fix(ludus): 🐛 refactor ludus provider for less code re-use and better version checking
